# public/api-documentation/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Laravel Swagger Documentation",
        "contact": {
            "email": "heningtonfrota@gmail.com"
        },
        "version": "1.0.0"
    },
    "paths": {
        "/api/register": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Cria um novo usuário",
                "operationId": "8e4e3cb7b46126f4696379d3c7eeb4ad",
                "requestBody": {
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "name",
                                    "password_confirmation"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Test"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "@123456789"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "example": "@123456789"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "User created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "User created successfully!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect fields",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The name has already been taken. (and 2 more errors)"
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/login": {
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Obter um token de usuário de autenticação",
                "description": "This endpoints return a new token user authentication for use on protected endpoints",
                "operationId": "222b2625838e3c57e017987f73598fda",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password",
                                    "device_name"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "test@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "@123456789"
                                    },
                                    "device_name": {
                                        "type": "string",
                                        "example": "APP"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Token generated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "plainTextToken": {
                                            "type": "string",
                                            "example": "2|MZEBxLy1zulPtND6brlf8GOPy57Q4DwYunlibXGj"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Incorrect credentials",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The provided credentials are incorrect."
                                        },
                                        "errors": {
                                            "type": "string",
                                            "example": "..."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/games/{id}": {
            "get": {
                "tags": [
                    "Jogos"
                ],
                "summary": "Obtém informações de um jogo específico",
                "operationId": "36b4a348cb892768bbe669a5e89ccc75",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do jogo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Informações do jogo",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Jogo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jogo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Jogos"
                ],
                "summary": "Atualiza um jogo existente",
                "operationId": "5fed7b0a0ab040d901dcf01b0b342d41",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do jogo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/JogoInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Jogo atualizado com sucesso",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Jogo"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Jogo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Jogos"
                ],
                "summary": "Exclui um jogo",
                "operationId": "a6179c952c5cff3a05b8777b33411f6f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do jogo",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Jogo excluído com sucesso"
                    },
                    "404": {
                        "description": "Jogo não encontrado"
                    }
                },
                "security": [
                    {
                        "bearerToken": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
            "Jogo": {
                "properties": {
                    "id": {
                        "type": "integer"
                    },
                    "nome": {
                        "type": "string"
                    },
                    "genero": {
                        "type": "string"
                    },
                    "preco": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            },
            "JogoInput": {
                "required": [
                    "nome",
                    "genero",
                    "preco"
                ],
                "properties": {
                    "nome": {
                        "type": "string"
                    },
                    "genero": {
                        "type": "string"
                    },
                    "preco": {
                        "type": "number",
                        "format": "float"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "bearerToken": {
                "type": "http",
                "description": "Acess token obtido na autenticação",
                "name": "Authorization",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}